+++ packages/applications/notifications/src/index.ts
@@ -1,2 +1,2 @@
-export * as AbandonNotification from './lauréat/abandon.notification';
-export * as GarantiesFinancièresNotification from './lauréat/garantiesFinancières.notification';
+export * as AbandonNotification from './subscribers/lauréat/abandon.notification';
+export * as GarantiesFinancièresNotification from './subscribers/lauréat/garantiesFinancières.notification';
