--- packages/applications/notifications/src/templateId.ts
@@ -1,21 +1,0 @@
-const abandonChangementStatutTemplateId = 5335914;
-
-export const templateId = {
-  abandon: {
-    accorder: abandonChangementStatutTemplateId,
-    annuler: abandonChangementStatutTemplateId,
-    confirmer: abandonChangementStatutTemplateId,
-    demander: abandonChangementStatutTemplateId,
-    demanderConfirmation: abandonChangementStatutTemplateId,
-    rejeter: abandonChangementStatutTemplateId,
-    demanderPreuveRecandidature: 5308275,
-  },
-  garantiesFinancières: {
-    dépôtSoumisPourDreal: 5740312,
-    dépôtSoumisPourPorteur: 5740394,
-    dépôtValidéPourPorteur: 5765321,
-    attestationGFActuellesEnregistréePourDreal: 5765406,
-    GFActuellesModifiéesPourPorteur: 5765519,
-    GFActuellesModifiéesPourDreal: 5765536,
-  },
-};
