+++ packages/applications/notifications/tsconfig.json
@@ -15,9 +15,6 @@
       "path": "../../infrastructure/pg-event-sourcing"
     },
     {
-      "path": "../../libraries/email-sender"
-    },
-    {
       "path": "../routes"
     }
   ]
