+++ packages/applications/projectors/src/index.ts
@@ -1,6 +1,6 @@
-export { seedAppelOffre } from './appel-offre/seed';
-export * as AbandonProjector from './lauréat/abandon.projector';
-export * as GestionnaireRéseauProjector from './réseau/gestionnaireRéseau.projector';
-export * as RaccordementProjector from './réseau/raccordement.projector';
-export * as TâcheProjector from './tâche/tâche.projector';
-export * as GarantiesFinancièreProjector from './lauréat/garantiesFinancières.projector';
+export { seedAppelOffre } from './subscribers/appel-offre/seed';
+export * as AbandonProjector from './subscribers/lauréat/abandon.projector';
+export * as GestionnaireRéseauProjector from './subscribers/réseau/gestionnaireRéseau.projector';
+export * as RaccordementProjector from './subscribers/réseau/raccordement.projector';
+export * as TâcheProjector from './subscribers/tâche/tâche.projector';
+export * as GarantiesFinancièreProjector from './subscribers/lauréat/garantiesFinancières.projector';
