+++ packages/applications/projectors/src/subscribers/réseau/raccordement.projector.ts
@@ -1,10 +1,10 @@
 import { Raccordement, GestionnaireRéseau } from '@potentiel-domain/reseau';
 import { Event, RebuildTriggered } from '@potentiel-infrastructure/pg-event-sourcing';
 import { Message, MessageHandler, mediator } from 'mediateur';
-import { removeProjection } from '../infrastructure/removeProjection';
+import { removeProjection } from '../../infrastructure/removeProjection';
 import { findProjection } from '@potentiel-infrastructure/pg-projections';
 import { Option } from '@potentiel-libraries/monads';
-import { upsertProjection } from '../infrastructure/upsertProjection';
+import { upsertProjection } from '../../infrastructure/upsertProjection';
 import { IdentifiantProjet } from '@potentiel-domain/common';
 import { getLogger } from '@potentiel-libraries/monitoring';
 import { CandidatureAdapter } from '@potentiel-infrastructure/domain-adapters';
